Move filesystem.glide to adapters.glide.cloudicon
