+++ app/src/main/java/com/amaze/filemanager/adapters/glide/cloudicon/CloudIconDataFetcher.java
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.amaze.filemanager.filesystem.glide;
+package com.amaze.filemanager.adapters.glide.cloudicon;
 
 import java.io.IOException;
 import java.io.InputStream;
